+++ src/pages/Admin/Charity/Banking/RecipientDetails/RecipientDetailsForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./RecipientDetailsForm";
