+++ src/pages/Admin/Charity/Banking/RecipientDetails/RecipientDetailsForm/useRecipientDetailsForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./useRecipientDetailsForm";
