--- src/pages/Admin/Charity/Banking/RecipientDetails/useForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useForm";
